requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==37.0.2
cssselect==1.1.0
channels==3.0.5
Django==4.0.5
django-cors-headers==3.13.0
django-multiselectfield==0.1.12
django-qr-code==3.1.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
lxml==4.9.1
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pyquery==1.4.3
pytz==2022.1
segno==1.5.2
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
dj-database-url==1.0.0
django-recaptcha==3.0.0